+++ src/components/FileDropArea.vue
@@ -1,16 +1,17 @@
 <script setup lang="ts">
 import { useI18n } from "vue-i18n";
+import { getFilesFromDataTransferItems } from "datatransfer-files-promise";
 import useMessenger from "@/message/messenger.composable";
 import useDropToPage from "@/components/droptopage.composable";
 
 const emit = defineEmits<{
-  (e: "drop", files: FileList): void;
+  (e: "drop", files: File[]): void;
 }>();
 
 const { alert } = useMessenger();
 const { t } = useI18n();
 
-function drop(event: DragEvent) {
+async function drop(event: DragEvent) {
   // On Chrome+Ubuntu, the file list may be empty for security reasons.
   // See https://askubuntu.com/a/1411727
   if (!event.dataTransfer?.files[0]) {
@@ -18,7 +19,9 @@ function drop(event: DragEvent) {
     return;
   }
 
-  emit("drop", event.dataTransfer.files);
+  const files = await getFilesFromDataTransferItems(event.dataTransfer.items);
+
+  emit("drop", files);
 }
 
 const { isDragover } = useDropToPage(drop);