+++ src/corpus/sources/sources.composable.ts
@@ -29,7 +29,7 @@ export default function useSources(corpusId: string) {
     return mink.downloadPlaintext(corpusId, source.name).catch(alertError);
   }
 
-  async function uploadSources(files: FileList, onProgress?: ProgressHandler) {
+  async function uploadSources(files: File[], onProgress?: ProgressHandler) {
     await mink.uploadSources(corpusId, files, onProgress);
     loadSources();
   }